refactored notification stuff into separate modules per context
